[breadboard] add base to context, disambiguate parent
invoke & co used the board in the context to compute the base url,
which meant
GraphToKitAdapter had to pass in the kit's board instead of the actual
board that is calling.

This had some issues:
- Integrity is based on the wires in the calling board and this lost that context
- Other properties such as kits that are passed down to the next board
   would come from the wrong board.

Also renamed #parent to #outerGraph to disambiguate from other
parent-child relations, e.g. nested board invocations.