[new syntax] remove unneeded type casting + helper
return values in `recipe` handlers often required `as unknown as string`
or so. if the output type of the node producing the output is specific,
this should no longer be needed. Where it's ambiguous (e.g. for
runJavascript), a new helper type V is exported so we can write
`as V<string>`.