[breadboard/new] don't use lambda nodes for non-closures (#237)
When passing a lambda that is a self-contained recipe, then no longer
add extra lambda nodes.

This allows `import { aRecipe } from "elsewhere"` and then using that
recipe, either as is or as lambda in a new board.

Note that if one imports an actual closure like this, then you are
creating a closure by using it and right now it'll serialize everything
that is connected. In practice there should be no need to include
closures like this (instead the entire closure should be wrapped in a
recipe), so this should probably produce an error on serialize. That is,
we allow serializing self-contained recipes that are defined outside of
the scope of the serialized recipe, but we don't allow including _nodes_
that are outside a serialized recipe. Note that the regular way to
serialize a closure is to run it and return it.