allow .title() and .description() on values, use in schema inference (#322)
In addition to .isString(), etc. this allows setting title and
descriptions of the schema.

This should now be enough to no longer need to pass a whole `schema` to
`input` and `output` and instead derive the schema from the connected
wires. Types are still quite basic, but it should be easy to add the
next set of needed features.