[breadboard-ui] Dedupe input/output/secret log entries (#346)
Having spoken with @dglazkov I realised that `nodestart` is the "key"
event through which we can drive the history log. In other words the
`input`, `output`, and `secret` events that were being created as a side
effect of the other UI work can be removed.

This PR makes a few changes:

1. It removes the any entry that isn't `nodestart|end` or
`graphstart|end`.
2. It uses the state object from the `nodestart` to populate an "input"
area in the entry's value field. (There isn't any corresponding state in
the `nodeend` for us to update the "output" side as yet.
3. It moves each `graphstart` event before its neighboring `nodestart`
event. Right now we emit the `graphstart` after the `nodestart` for a
subgraph, and that makes things look a bit peculiar; intuitively I think
you want to see `graphstart` and `graphend` "containing" the relevant
nodes.

Number 3 is a little hacky, therefore, as I feel like we'd be better
placed doing that work in the harness rather than the UI.