[breadboard-ui] Dedupe types (#347)
This PR simplifies the types usage; we now take the message as-is from
the harness to the UI and then change it around a little for
presentation. This has allowed me to remove virtually all of the
`HistoryEvent` types, which was mostly duplicative and
very-similar-but-ever-so-slightly-different to the main Breadboard state
types.

In the next PR I hope to simplify things a little further.