[breadboard-ui] Include last event time in updates (#354)
I noticed that we don't update the last event time in the
`#updateHistoryEntry` function, which we should. This means that we
don't get live updates to the history in every case. We also need to key
the JSON view off the last updated time so that Lit knows to re-render.