[breadboard-web/-ui] Move UI components to web (#415)
This is a bit of moving around because I was finding that trying to roll
the Preview into the UI controller was making things odd. The files for
the Preview need to live as siblings to the UI Controller, but the
existing approach required them to be available _inside_ the UI
Controller.

What I've done here, then, is to move the board list and "outer UI" up
to breadboard-web, and breadboard-ui's UI Controller now takes care of
rendering the debugger alone. In the future I think we'll probably see
the Preview use breadbuddy or a playground, but for now I'm building
something inside of breadboard-web. And I suspect UI Controller needs to
be renamed to `bb-debugger`.