[breadboard-cli] Use debugger/web as a dependency (#434)
We currently copy in the `dist` from breadboard-web into `dist/ui` so
that we can run the debugger. This PR changes things so that we now have
an additional vite-backed build that generates `dist/debugger`. I've
created three entry points: `index.html`, `preview.html` and `worker.ts`
and, inside each, I pull in the parts of `breadboard-web` that are
needed.

This allows us to treat `breadboard-web` as an "out-of-tree" dependency
if we so desire.