[breadboard-web] Fix duplicate UI (#657)
Fixes #637 

We were restarting the harness but not creating a new UI when switching
between Build & Preview. This caused the UI to receive a new set of
messages from the new harness which, in most cases, would be duplicates
of the original harness. And in turn that caused duplicate inputs as
logged in #637.

This PR fixes it by tracking the board ID so that we avoid starting a
new harness if we've not changed boards. The net effect is that Build
holds onto the current state of the world whereas Preview is a fresh
copy every time we switch to it. I _think_ that's the right behavior
here, because intuitively going to the Preview feels like something you
do when you're wanting to run through a journey from the start... but I
can change that if needed.