[breadboard] Allow pre-existing runners (#700)
One of the challenges I've run into with visual editing is that our
board runner expects a URL. In the case where we're modifying a board
"in memory", we don't have a JSON board to load via fetch. What I've
done here, which might not be the right way (in which case we can just
close this PR), is to allow a `BoardRunner` in the config. If that's set
it's used in preference to the URL for `run`. This just means I can do
something like the following:

```ts
const runner = await BoardRunner.fromGraphDescriptor(
  this.loadInfo.graphDescriptor
);

const config = createRunConfig(this.loadInfo.graphDescriptor.url);
config.remote = false;
config.proxy = [];
config.runner = runner;
```

And now I can run a modified board locally. @dglazkov is there a better
way to do this?