[core-kit] Add resolve node (#814)
Adds a `resolve` node to `core-kit`.

It expects inputs that are relative URL strings, and returns absolute
URL strings on outputs with the same name.

The special `$base` input sets the base for resolution. By default, it
is the URL of the graph containing the node itself.

This is useful for addressing the problem raised in
https://github.com/breadboard-ai/breadboard/issues/733, where if you
have a relative path (e.g. a user input specifying a board URL to load),
and you pass it to a child graph which has a different base URL to the
parent graph, and then pass that path to invoke, invoke will try to
resolve that path relative to the *child* graph instead of the *parent*
graph, thus getting the wrong URL. The solution to this scenario is that
the parent graph should first pass the user's path through a resolve
node before passing it down to the child, so that it gets resolved with
the correct base path.