Implement and test type system for monomorphic definitions (#1022)
Implement the type system for `defineNodeType` in the new
`@breadboard-ai/build` package.

Example usage:

```ts
import {defineNodeType} from '@breadboard-ai/build';

export const reverseString = defineNodeType(
  // Inputs
  {
    forwards: {
      type: "string"
    }
  },
  // Outputs
  {
    backwards: {
      type: "string"
    }
  },
  // Invoke function
  ({forwards}) => {
    return {
      backwards: forwards.split("").reverse().join("")
    }
  }
);
```

This gives you a node definition which can be invoked directly as a
function to create a node instance.

The types are enforced strictly. The invoke function must match the
types of the declared inputs and outputs. An instance will have input
and output ports that are parameterized with the declared types.

Part of https://github.com/breadboard-ai/breadboard/issues/1006