[breadboard-ui] Fix case sensitivity of file URLs (#1109)
I noticed that files would fail to load from a mounted a folder with
capital letters in its name. Because URL creation converts to lowercase
it meant that we were failing to match the location portion of the URL
with the folder name. Now we always store the location as lowercase, but
we query the handle's name for the title in the nav.