[breadboard-ui] Check for icons (#1116)
I was assuming that if `visual` existed on `metadata` that we were
dealing with a node which had an icon. In fact it may just be that the
node has `[x, y]` coordinates. This PR changes things so we confirm that
there is an icon before we set the class and icon value in Activity Log