[breadboard-web] Enforce lowercase entry names (#1220)
I noticed that FS API labels can have capitals, but because we use URLs
(which generally converts components to lower case) we were failing to
match. One option was to do a second pass with case insensitivity when
we fail to find a match between the URL and the FS API label, but in the
end I opted for lowercasing the file name. This will have the
corresponding effect of erroneously deduping FS API labels when files
share the name but are differentiated by casing, e.g., foo.json and
FOO.json, but hopefully that won't occur often enough to be problematic.