[build] Require all concrete values to be JSON serializable (#1231)
This adds a new `JsonSerializable` type which is now the base type that
all `BreadboardType`'s must be convertible to. This e.g. eliminates
`undefined` as a possible value, since that's not JSON serializable. We
want everything to be JSON serializable so that we can easily support
save/restore for runs.