[build] Parameterize ports by actual type instead of port config type (#1232)
This simplifies IDE tooltips a bunch, e.g. instead of `InputPort<{ type:
"string"; }>` we now will just show `InputPort<string>`.

It also fixes an assignment bug in the type system, since we were
previously asking the type system to understand that `{type: "string"}`
was assignable to `{ type: anyOf("string", "number") }`, which it
doesn't actually know; we have to convert to actual types first (since
it does know that `string` is assignable to `string | number`).