[breadboard-ui] Update secrets in UI (#1339)
Fixes #1335, but in a different way. We were capturing the settings, but
not requesting a new copy of them on update. The "fix", small as it is,
is to call `requestUpdate` on `<bb-main>`, which will re-run `render`
and which will, in turn, fetch a new copy of the settings and pass them
through to the UI.