[core-kit] Add `code` function (#1621)
The `code` function creates a `runJavascript` Breadboard node in a
type-safe manner. Usage:

```ts
import {code} from "@google-labs/core-kit";
import {board, input} from "@breadboard-ai/build";

const forwards = input({ description: "A string to reverse" });

const reversed = code({ forwards }, { reversed: "string" }, ({ forwards }) => ({
  reversed: str.split("").reverse().join(""),
}));

export board({forwards}, {reversed: reversed.outputs.reversed});
```

As a future improvement, we could make the `outputs` parameter optional,
but we'd
have to be careful with schema. What we'd need to do if `outputs` is
missing is:
  
1. Infer the output ports from `ReturnType<fn>` instead of from
`outputs`.
  
2. Return a node that has a magic `outputs` object. This `outputs`
object
would be a proxy that overloads `get` to automatically create output
ports as they are accessed (same as what we do below, except deferred).
  
3. Deal with the fact that the runtime `type` of those ports will be too
broad. It will be the base dynamic output type of `runJavascript`, which
     is any JSON serializable value.
  
This is OK when wiring to an input port because input ports already know
their schema, and TypeScript will take care of making sure we only wire
     it to the right place.
  
However, if we want to use one of these outputs as a _board output_, we
have a problem, since we need to write the schema to the output node
     configuration. So, we'll probably want to introduce a "weak typed"
     version of an `OutputPortReference`, which is accepted when wiring
nodes, but NOT accepted when configuring board outputs. Then, we can
     update the `output` utility function to only take weak
     `OutputPortReferences` if the `output` has had its type explicitly
     specified.