[build] Pass inputSchema and outputSchema to describe functions (#1677)
`describe` functions are now passed `inputSchema` and `outputSchema`.
Unlike the previous `describe` signature, though, now those two
properties are part of the `context`. This just makes the signature
easier a bit easier to read, instead of passing those properties as
their own parameters.

The schemas are converted to `PortConfig` objects, which are very
similar, but use Breadboard Type Expressions as the `type` field,
instead of JSON. The original schema is bundled into a special
`unsafeType` wrapper so that it's preserved exactly if the
describe-author decides to return the incoming schema unchanged.