[breadboard] Clarify meaning of required output ports (#1813)
Before this PR, effectively saying that an output port was required
meant "you must wire this up", since it would turn red in the UI if you
didn't. However, really if an output port is required, it's just saying
that it won't be undefined, and it has no bearing on how the caller
choses to use it or not.