[breadboard-web] Fix specialist testing grounds (#1847)
I noticed we had an array of LLM Contents as the examples for the
specialist, i.e., examples is an array and the zeroth item itself was an
array of LLM Contents... whereas the input itself is set to a single LLM
Content. This meant that the LLM Input received an array when it
expected an object.