[connections] Adds a Connections settings panel (#1978)
Adds a "Connections" settings panel which talks to the Breadboard
Connections Server and displays the list of available connections. It is
not yet possible to actually *connect*, that will come in followups.

<img
src="https://github.com/breadboard-ai/breadboard/assets/48894/a06341c7-c70a-4dbe-ad10-48bc2e3fa1ff"
height="250px">

<hr>

I added two general-purpose features to facilitate this:

1. You can now render a completely custom settings panel UI. The
connections panel is quite specialized, so it's nice to have full
control over the UI instead of using the generic settings form/widgets.

    ```js
    [BreadboardUI.Types.SETTINGS_TYPE.CONNECTIONS]: {
      configuration: {
        extensible: false,
        description:
"Third-party services boards can access. When you are signed into a
service, any board can access and modify your data on that service.",
        nameEditable: false,
        nameVisible: false,
        customElement: "bb-connection-settings",
      }
    ```

2. Added an `Environment` object which can be consumed by any element
under `bb-main` like this:
   
   ```js
   @consume({ context: environmentContext })
   environment?: Environment;
   ```

The idea is for `Environment` to contain deployment-environment-specific
configuration. In this case, I need the URL of a Breadboard Connection
Server:
  
    ```js
    export interface Environment {
      connectionServerUrl?: string;
    }
    ```
  
Currently this value is provided by `bb-main` with some hard-coded logic
that will work for local development (if we're on localhost:5173, then
return `localhost:5555` as the connection server).