[Debugger] Fix store provider for nav (#1984)
Fixes #1977

The faulty assumption on my part was that the IDB provider would always
be available. In the debugger case it isn't, so where we fail to load
the default provider we switch it across to whatever provider is
available. I've been a bit restrictive here inasmuch as I only perform
that switch when there's only one provider available, but we can relax
that constraint if it proves problematic.