[connections] Hook up connections to the secrets system (#2001)
If a `secrets` node requests a secret starting with the prefix
`connection:`, then we will prompt them to sign in to the corresponding
service instead of showing the normal secrets collection UI.

If no matching connection is found on the configured Breadboard
Connections Service, it will show "Could not find a connection for foo".

If there is already a connections entry stored in settings (i.e. because
they signed in earlier, either from another board run or directly from
the settings panel), we check if it's expired or will expire within a
minute. If so, we automatically refresh it. If not, we use it right
away.