[inspector] Add a validate method to InspectableEdge (#2345)
Adds a `validate` method to `InspectableEdge`. It grabs the schema for
the from/to sides of the edge and calls `canConnect` to determine
whether they are compatible. Includes a new Validation type so that we
can add more structured data, detailed error messages, etc.

Part of https://github.com/breadboard-ai/breadboard/issues/2298