[breadboard-web] Unblock UI on providers (#2352)
We were waiting on board lists being restored before rendering the UI,
which was slowing things down in the case of Board Servers particularly.
This PR changes things so it eagerly renders while the Providers are
restoring. To accommodate that change the main loader awaits the
selected Provider's `ready` `Promise` before loading a board, and the
nav similarly awaits the same before trying to render the item list.