[visual-editor] Fix nav filter and delete behaviors (#2525)
I noticed that in fixing the nav scroll issue (which was essentially
being overly keen to re-render the nav) we are now too keen in the other
direction. This means that when a board is created or deleted, or when a
provider is updated, we don't re-render the nav (even though we do
perform the action).

This PR increases the set of things that will trigger a re-render of the
nav.