[data-store] Introduce Run Store (#2672)
This is a fairly hefty change.

1. Runs are still created and managed in the `RunObserver`.
2. As results are added they are added to the current run _and_ sent
over to the `RunStore` instance (if set in the config).
3. The `RunStore` instance is behind a factory function which will
choose which `RunStore` to instantiate based on capability. The default
`RunStore`, included in `packages/breadboard`, is an in-memory one which
works for both nodejs and browsers. Browsers will likely receive an
IDB-backed one from `packages/data-store`.
4. If a `RunStore` is configured, runs are restored from it when the
`RunObserver` boots up. (This will only be useful for `RunStore`s with
permanence, like IDB, File System, etc).
5. As results are added to a run they are also checked for `inlineData`
parts, which are converted to `storedData` instead via the `DataStorage`
instance (again, set in the `RunObserver` config). If there is a
`storedData` part in the run, it is assumed to be from a _previous_ run
and its value is copied for the current run. This prevents the case
where we lose blobs after a couple of runs because the `storedData` was
removed.
6. If a `storedData` part is sent over to a `RunStore` it is assumed
that the `RunStore` will inflate it to an `inlineData` again _before_ it
stores the data, that way we don't persist data that can't be restored
later on.
7. Each `Run` now has a unique ID, which is used to key into the
`DataStorage` (rather than the old group ID number).
8. The `DataStorage` instance is a more fine-grained than before
inasmuch as it uses the `Run`'s unique ID, as well as the node's
timestamp and properties to identify the part. This allows it to
selectively deflate parts it hasn't seen before.