[board-server] Create a ServerConfig object (#2901)
For now, the config contains the value already being used by the server
to route
and serve requests. It wille eventually contain allowed origins (just an
empty
set for now).

Pass the config object down to whatever handlers require it. As a
followup, we
may also want to revisit this routing logic to handle config-specific
tasks
before delegating.