[board-server] Read CORS config from environment variables instead of Firestore (#2918)
Part of #2896

Switches from reading the Firestore table to determine CORS headers to
using
the configured environment variables.

THIS IS A BREAKING CHANGE. All existing Board Server deployments will
need to
have their environment config updated in order to continue functioning.

In the spirit of making the breaking change as minimal as possible, this
PR does not delete the code that reads the (now redundant) Firestore
table. That will be done in a separate cleanup step.