[all] [dev] Make linting depend on tsc (#2935)
There was some bad lint caching behavior happening, because each
package's lint command was only checking whether its own package files
changed. We need to actually build the typescript of all dependencies
first, or else when the linter reads the types of those deps, they might
be out of sync. Depending on `tsc:build` is the easier way to do that.