[build] Set input port to empty string when wiring star ports (#2969)
If the input port is also "*", then the runtime seems to just stop. Only
"" works. However, if you paste some BGL with "" into the Visual Editor
it errors, and the docs say that star ports "can only be wired to other
star ports". Seems a bit odd, but I'll go with the one that works in the
runtime, which is empty string.