[shared-ui] Fix layout reset bug (#3005)
I noticed that comments no drag as part of a group after a layout reset.
It turns out we were removing all the layout information but only
restoring the the non-comment nodes with dagre. Since the comments have
no edges we can't use dagre, but we can just restore their existing
layout info, which is what this PR does.