[build] Serialize nested components with their kit-name if they are instantiated via a kit (fixed version) (#3030)
Brings back https://github.com/breadboard-ai/breadboard/pull/3015
(reverted in https://github.com/breadboard-ai/breadboard/pull/3018) but
with a fix for a bug.

The bug was that when building the kit, I was using the "id" property of
the component object itself as the key, instead of the key that was used
to initialize the component within the kit. Combined with a removed get
proxy, this meant the key was either undefined, or potentially a
different id from how the kit was initialized. Added tests and some
safer typing around this.