[Breadboard] Move the provision of the default graph provider (#3316)
Part of #3314 

This is a no-op change, since this is the only caller of the constructor
for `Loader`

Add the DefaultGraphProvider in createLoader instead of the constructor
for Loader. This allows us to add logic in `createLoader` which
conditionally adds the default provider.

We're not copying `providers` in the constructor because `createLoader`
is its only caller. It should be safe to store the original array
without copying.