[code-build] Emit a legacy script instead of module (#3324)
runJavascript doesn't support javascript modules, so we'll have to emit
a legacy script for now. Doing this by asking esbuild for an iife. We
also need to strip the iife wrapper from that output, because
runJavascript also has some regexp logic (convertToNamedFunction) which
I think mis-parses the iife wrapper and breaks the code.

Also adds a test that the generated boards are truly executable!