[board-server] #3172 refactor board server to use express (#3340)
Does not *fix* #3172 as I feel this is still relevant until this is a
bit more mature. But this is the first pass of moving everything to
express, and since this lives in a more or less isolated tree, I think
we are approaching a good point to merge this stuff. This express
version passes the board server integration tests.

---------

Co-authored-by: Bertie Spell <berniesnell@gmail.com>
Co-authored-by: tina.kingsley-hughes <tina.kingsley-hughes@sourcepulp.com>