Basic setup of Wasm-based JavaScript sandbox (#3611)
So sadly it wasn't quite as straight forward as I hoped, but here is a
working setup to compile QuickJS to Wasm and use it as a sandbox.

To outline some of the adjustments that I had to make:
- Couldn't use wasm-pack, as wasm-pack targets `wasm32-unknown-unknown`,
but since this pulls in C code, I need to target `wasm32-wasi`. Luckily,
wasm-pack is really just a glorified script that runs multiple tools, so
wrote a build script that does mostly the same things.
- For the same reason with the target triplet, stock wasm-bindgen does
not work. I created a fork that makes wasm-bindgen work the same way for
`wasm32-wasi` as it does for `wasm32-unknown-unknown`. See
[commit](https://github.com/surma-dump/wasm-bindgen/commit/9c48bd37cd210878d44bda6c13e198d64e84c3d5)
here.

The rest is hopefully not too surprising. The brotli'd WebAssembly
binary comes it at around 300kB right now.