[connection-server] Accept only GET requests
Previously connection server was accepting all requests, but setting
Access-Control-Allowed-Methods to "GET". So it's clear that it's
intended only to be used with GET requests.

Arguably, however, `grant` and `refresh` requests should be POST
requests. They don't alter state on the connection server itself, but
they do cause outgoing POST requests, so it makes sense.

Also for `grant` requests specifically, the params include a refresh
token, which is a sensitive value that should not appear in the search
query of a URL. In the POST body it would be encrypted, and less
vulnerable to sniffing attacks.