[bbrt] Formalize Driver as the interface that drives the main conversation (#3910)
Introduces a `Driver` interface: an object that executes a turn (i.e. it
provides the connection to the language model). For now, this just needs
a name, an icon, and the ability to execute a turn. It is easier now to
see how a Breadboard itself could be a Driver.