[unified-server] Add visual editor (#4128)
Part of #3913

Committing the latest round of changes in the unified server to keep it
up to date.

This adds the Visual Editor as the root landing page of the unified
server. However, it doesn't work yet.

Specifically:

1. The client-side build ("vite build") clobbers the server-side build
("tsc") when it's run after. I'm not sure how to prevent this. If you
manually run the server-side build after the client-side build, then it
works, but wireit seems to want to run it in the opposite order.

2. The visual editor doesn't actually load. Instead, a blank page is
shown. The error in the console says that it's attempting to load
"/core-kit.js", which it reads as a string from the "kits.json" file.
Since this isn't build by Vite, it doesn't serve. There may be another
staging step required.