[unified-server] Remove tests from global test run
This works around the error seen in #4249 by removing the tests from the
CI run.

This is acceptable for now because the current unified server tests are
just dummies that "assert true" to verify that they are run. We should
add these tests back to the global run when #4249 is fixed.