[google-drive-kit] Support images in append-to-doc (#4256)
Images are now included in the docs created by `append-to-doc`. It works
by uploading an image to drive, making it public, and then deleting it
(copied from the existing logic from @dglazkov in `context-to-slides`).

Fixes https://github.com/breadboard-ai/breadboard/issues/4223, with
caveat that it does not seem to work on corp due to restrictions around
creating public file links. Filed
https://github.com/breadboard-ai/breadboard/issues/4257 to track that.