[shared-ui] Show updating status on components (#4348)
This PR sets components to be gr(e|a)y when the `updating` flag is
`true` and it hides their ports. It's probably something we want to
revisit, but for now it does delineate the component from its loaded
state pretty well.

Also fixes a bug where the icon doesn't show up in the port label.