[unified-server] Stage index.html during build
This change replaces the hand-coded version of the `index.html` (which
was copied from visual editor) with a version that is copied
automatically during staging, as a prerequisite to the build.

Entry points do not change.

The only meaningful difference between the two HTML files was that the
unified server version replaced this:

https://github.com/breadboard-ai/breadboard/blob/c3816e4a0dd81894693b174a70050151260e0303/packages/visual-editor/index.html#L42

With this:

https://github.com/breadboard-ai/breadboard/blob/c3816e4a0dd81894693b174a70050151260e0303/packages/unified-server/index.html#L38

However, a direct copy from the visual editor works as long as the
`Main` class can be imported from a source file found at
`./src/index.ts`. So a file is added to unified-server that re-exports
the the visual-editor module.