[unified-server] Remove unneeded dependencies (#4430)
`docker-build` does not need `build` because the build is run when the
image is built.

Forcing `docker-run` to depend on `docker-build` can take time when the
caller might not want to build a new image. The user should just run
`docker-build` and `docker-run` in sequence if they want to build a new
image. As written, `docker-run` will just take whatever image is tagged
with `unified-server:latest`.

Part of #4355