[board-server unified-server] Refactor board server base to use Express (#4585)
This change reworks the initialization logic for board server so that it
is created as an Express app, even in the board-server package. This
gets rid of the need to export `makeRouter` from the board-server
package, so now only the single module is exported. The package.json is
updated accordingly.

Everything needed outside of the board server package should be exported
by the `server` module, which is the default export of the package. At
the moment this includes `createServer`, `createServerConfig`, and the
config type.

Part of #3172 and #4353