[board-server] Add a skeleton integration test (#4634)
This is a basic test that simply tests the routing logic for board
server. It only asserts on the response status code, and does not
inspect responses, nor check for side effects.

This is a basic level of test coverage needed to get some confidence in
an ExpressJS refactor of the routing logic.

The process of writing this test exposed a number of bugs and confusing
behaviors. Specifically, the SQLite storage provider doesn't seem to
work as expected. This test can be used as a basis for getting it up and
running again.

Part of #3172