[unified-server] Tease apart dependencies for faster reload loop (#4642)
Previously, the dev-mode unified-server would restart whenever any
client-side code changed. Now, only changes to the implementation of the
unified-server itself trigger a restart, not client-side changes. So
client-side changes should get picked up more quickly with less page
reloading.

The actual change here is to switch the server-to-static-assets
dependencies from being regular wireit dependencies (which will cause a
server to restart on any change), to non-cascading dependencies (which
will keep that dependency fresh, but not restart on changes -- see also
https://github.com/google/wireit?tab=readme-ov-file#execution-cascade).

And in order to do that, we also needed to split apart the client from
server code a bit in the build graph, including 2 separate tsconfigs.