[board-server] Use res.locals to store board info (#4670)
Store the board info, extracted from the path, on the Response.locals
object so that it doesn't need to be passed down to handler functions.

The goal is to eliminate as much logic as possible from boards/index, so
that it's really just a router. This doesn't do the whole job, but is a
step forward.