[board-server] Use express.json (#4704)
This is a partial re-revert of #4688 (reverted in #4699)

The original PR caused proxy requests to 400 because the middleware was
not setting the `Request.body` attribute.

This is because the request to the proxy endpoint has `Content-Type:
text/plain;charset=UTF-8`, so the Express middleware is not recognizing
it as a JSON request.

This PR works around the issue by not using the middleware for that
request. But the request body is JSON, so if we set the request header
correctly, we could use the standard middleware for this endpoint.