[board-server] Unlink SQLite storage provider (#4782)
Gets rid of the generic BoardServerStore interface and replaces it
explicitly with the Firestore version. This allows us to change the
Firestore version without breaking the build or updating the SQLite
version.

This is hopefully a temporary change. Have filed #4781 to get this
hooked back up. But the whole storage layer will have changed in the
meantime, so there will likely be some manual work to do to get it
working again.