[board-server] Require current user to get board (#4962)
Previously we were only checking for board permissions in the `GET`
handler(s) for the boards API. This change pushes that auth check down
into the storage layer.

Prior to this change, the invoke handler could actually be used as a
workaround for auth checks. Since it would load any board by name that
was embedded in the owned board. As long as the initial board could be
loaded, all others would succeed without checking permissions. This
would allow private boards to be invoked via the invoked handler without
permission.

Overall, we *need* to push the auth logic down into the storage layer,
since a lot of our auth logic is specific to the storage mechanism.
Different storage providers will have different ways to do auth.

Part of #4960