[revision history] Remove pause/resume (#5201)
The pause/resume feature was a way of consolidating contiguous related
edits, but since it was keyed on label, it behaves in a slightly
unpredictable way (because labels are themselves unpredictable;
sometimes UUIDs, sometimes human-readable).

I'd like to add-back a related-edit consolidation feature, but I think
it's more likely to be time-based (or time + node ID or something)
rather than label-based.

In the meantime pause/resume is making it a bit tricky to reason about
the state of history, so I'm pulling it out for now.