[Drive] Dedicated app view requires login for Drive graphs (#5725)
Also includes a fix to the share panel to allow it to display the app
link for graphs not owned by the current user.

Note there are still some issues where the graph seems to have trouble
executing, like an OAuth issue, but the same thing is happening with our
Firebase-backed version too, so I'm treating that as a separate issue.