Use EmbedHandler to support iterate-on-prompt button, homepage loaded, and creating a new board message. (#5781)
Implements the following three features in communicating between an
iframed breadboard page and its embedder:

1. **Iterate on prompt**. Allows the embedder to send an
`toggle_iterate_on_prompt` message which enables a new button in the
title of the expanded node. Pressing this button emits a
`iterate_on_prompt` message to the embedder conveying information about
the board and node of origin, as well as the prompt at the time of
button press.

1. **Create a new board.** Allows the parent to send a
`create_new_board` message. If the requested board is succesfully
created, the generated board ID is sent back as a message to the
embedder.

1. **Home loaded.** Sends the embedder a message notifying it that
Breadboard loaded, along with the user's signed-in status.

1. **Handshake ready.** Sends the embedder a message notifying it that
Breadboard is ready to receive messages. This is necessary to prevent
the embedder from sending messages before Breadboard has set up the
necessary event listeners. (In this code, we do not receive the
`handshake_complete` message, as there is nothing that needs to happen
then.)

Also renames the class of message sent by the embedder as
`EmbedderMessage` to distinguish them from `BreadboardMessage`, which
denote messages sent by Breadboard.