Retrieve & validate oauth redirect URI provided by embedder (#5823)
To support auth in embedded pages, retrieve and validate OAuth redirect
provided as a URL param.

For the oauth redirect to work:
- The deployment must specify the environment variable
`VITE_VALID_REDIRECT_URI_ORIGINS` with a JSON list of valid origins for
the redirect URI.
- The embedder must pass in a `oauth_redirect` param, whose origin must
included in `VITE_VALID_REDIRECT_URI_ORIGINS`.

Additionally, this change also sends a message to the embedder conveying
whether the access request has succeeded or failed.