Correct board URL logic and declare interfaces for Breadboard-embedder communication (#5842)
Small fixes for breadboard-embedder integration:

- Corrects logic to extract board URL/ID. The prior code treated
`graph.raw().url` as an HTTP/HTTPS URL, where the "board ID" was the
`tab0` search param within that URL. In actuality, the `graph.raw().url`
is the board ID (or rather, board URL). If this is not available; then
we extract the `tab0` value from the current window.location.
- Add "declare" for all interfaces shared between breadboard and
embedder to ensure field don't get renamed by the compiler. This is not
currently a problem, but as a precaution, making it explicit.