[embed] Handle empty Create-new-board message & send message when home logo pressed (#5845)
Currently, Breadboard will throw an error if an embedder passes a
`create_new_board` message with no prompt. With this change, that in
should create an empty board that the user can populate.

Additionally, this sends the "back_clicked" message when the user clicks
on the homepage logo from within a board.