[sharing] Add special escape key race handling for granular sharing dialog (#6019)
If the user clicks "View Share Permissions" and then very presses
"Escape" before the granular sharing dialog (which is a component that
we don't control that provides no events relating to opening and
closing, so we have to resort to a weird MutationObserver hack to figure
out what it's doing), we previously would get in a weird state where the
dialog never loads, but we're stuck waiting for it. Now we add a special
case just for this.