[sharing] Invalidate user list cache so that shareable copies don't appear in the gallery (#6021)
This shouldn't be necessary, but if we don't do this, the copy will
appear on the user's gallery page until the browser's site data is
cleared because of our Drive caching infrastructure. This is a quick
hack in lieu of understanding how to make the caching infrastructure
handle this case better.