Fix bug connecting to Google Drive from embedded page. (#6054)
Allows users to upload Drive assets when Breadboard embedded in another
page.

The `PickerBuilder` requires the origin to be set to the top-most origin
when the application is running in an iframe. (See
[docs](https://developers.google.com/workspace/drive/picker/reference/picker.pickerbuilder.setorigin.md).)
Hence, we retrieve the origin from the query param already provided by
the embedder, and as usual vet it against the list of valid embedder
origins. If valid, we use it to set the `PickerBuilder` origin;
otherwise we use the standard window origin.