[gdrive] Maybe we should refresh the UI immediately when a gallery item gets deleted? (#6077)
So far we were only refreshing if the item changed because of the
actions of the user themselves.

IMO there is a drawback of what I'm doing here since an item can
disappear from under the users cursor or UI can refresh and jump without
user doing anyting.

Also IMO a more proper way to implement this is to gray out the
[deleted] element and make it unclickable while the UI remaining stable,
with a snack bar saying "list changed, click here to refresh". However
this is too much effort for the moment so I'm just sending refresh event
in this PR.

Co-authored-by: Volodya <volodya@google.com>