[devx] Fix double reload for most edits by moving ClientDeploymentConfiguration to types package (#6230)
The double reload was happening because the unified server itself (the
node code) was depending on shared-ui, in order to import the
ClientDeploymentConfiguration type.

That caused the server to restart for any shared-ui change, which was
almost always unnecessary, and which caused the double reload (once when
the client code changed and the first instance of the server noticed,
and then again as the second instance of the server replaced the first
one).

So, I moved ClientDeploymentConfiguration to the types package, and
removed the wireit dependency between the server and shared-ui. The
double reload will now only happen when modifying something in the types
package, which should be much less frequent than shared-ui.