[init] Fix bug where on first load galleries would be empty (#6292)
This is caused by the way we try and coordinate board servers via IDB.
On the very first sign in, you won't have any IDB state, and we did not
have a mechanism for updating the main page once the initialization of
the Drive server finishes. That meant on first load, we rendered the
home page as though there were no board servers, and then never update,
so nothing would work until you reload the whole page again.