[sharing] Overhaul of asset sharing behavior (#6331)
- Assets are now either _managed_ or _unmanaged_.

- _Managed_ assets include theme splash images and any asset uploaded
directly through the upload UI. They are assumed to be owned by the
owner of the graph. Managed asset permissions are automatically
synchronized with the main graph (i.e. whenever a sharing permission is
added or removed from the graph, the same is added or removed from all
managed assets with no user interaction).

- _Unmanaged_ assets are assets selected with the Drive picker.
Unmanaged asset permissions are **not** automatically synchronized with
the main graph. Whenever a sharing permission is added or removed from
the graph, a dialog will now pop up asking the user how they want to
deal with the unmanaged asset(s). If an unmanaged asset is editable by
the user, and is missing a permission that the graph has, this is
explained and a "Share my assets" button is available to synchronize
them. If an unmanaged asset is **not** editable by the user, we simply
warn the user that it might not be available to all users, and suggest
they contact the owner to check, or replace it to be sure.