[shared-ui] Prioritize visible floating inputs (#6398)
I noticed that we don't reliably focus the floating input. This is
because we aren't differentiating between the three instances we have:
the console view, the preview view, and the app view. This PR
distributes the current UI state so that the floating input can compare
the in-focus view with its own configuration to decide whether or not it
should attempt to focus the `textarea` it contains.