[signin] Check for required scopes and force signin when missing (#6430)
We will now check that all required OAuth scopes are granted at
bootstrap, and if one is missing we redirect the user to the "Additional
access required" screen on the landing page.

We also now store scopes in the local signin storage. If, during the
missing scopes check, we find that we don't have any scopes in local
signin (because it is a signin from before this PR), then we call the
https://oauth2.googleapis.com/tokeninfo endpoint to read the scopes, and
automatically upgrade the local storage for next time.