[various] Track opened versions for shared apps (#6492)
To this point a user doesn't know if an app they've opened has been
updated by the author. This PR includes a quality-of-life improvement
whereby the version of an opened app is retained in browser storage. If
and when that same app is opened in future the version number is checked
and, should the app have been updated, the user is shown a snackbar
telling them as much.