[urls] Consolidate URL creation and parsing (#6497)
Adds new makeUrl and parseUrl functions which understand all of the
kinds of URLs in our app, and updated every place (at least those I
could find; there might be a few stragglers) where we create or parse
URLs to use this.

One bug this fixes is that previously if you were signed-out but clicked
on a link with a results file, the results parameter got dropped.

In general this should make it easier to create valid consistent URLs,
and will make it easier to change our URL scheme if we want, too.

Also adds a test script to the shared-ui package. It's a node test, but
that's the simplest/fastest way to run unit tests assuming it's not
depending on any web-specific apis.