[su/ve] Improve behavior of Remix button (#6510)
There is a little flakiness / inconsistency when using the Remix button.
This PR neatens things up in a few ways:

1. It enables the `inert` attribute on the header whenever there is a
blocking action. This means that when there's something blocking (like
remixing a board) it should now prevent multiple clicks of the button.
2. The checks in the remix button are slightly more relaxed since the
`inert` takes place of the former behavior that disabled the button when
clicked (also in an attempt to avoid multiple clicks).
3. We show the remix acknowledgment right at the start of the handler
before any work begins so that the user is assured that the action has
begun.