[shared-ui] Enable experimental 2D matrix rendering of graph (#6519)
By default we use 3D matrices for rendering nodes & edges which creates
compositor layers in Chromium-based browsers. For larger graphs
particularly this can create memory pressure where the tab needs more
memory than it can access, which in turn causes rendering bugs.

This PR enables an experimental flag for using 2D matrices (and
bypassing `will-change`) on edges and nodes. The trade-off is this:

1. With this flag off (the default) everything is given its own
compositor layer. This means we avoid re-rasterizing but for larger
graphs we can see memory pressure/exhaustion.
2. With this flag on everything is drawn into the same compositor layer.
This means that every pan & zoom action will cause re-rasterizing, but
on the upside everything is on a single compositor layer so there's less
to manage at that level.