[shared-ui] Capture pointer when renderer clicked (#6575)
When autoname starts we seem to force the focus away from the renderer.
This means we end up with a hanging drag behavior. The fix, simple to
write, but tricky to find, is that we need to capture the pointer when
we know that the user has clicked on the renderer. That way we get the
corresponding pointerup event, even if autoname has started.