[signin] Support incremental oauth scopes (#6584)
`SigninAdapter.token()` now takes an optional `scopes` argument, which
is an array of scopes defined in the new
`connection-client/src/oauth-scopes.ts` module. If `scopes` is set, and
one or more of those scopes have not been granted by the user, a new
dialog will popup asking the user to add that permission by going
through the OAuth flow.