[connection-server] Allow loading client secret from file (#6625)
Adds a property to the secrets.json file that specifies the location on
disk of the client secret, rather than the secret itself.

Since the client secret is the only sensitive value in the config, this
can allow secrets.json files to be non-sensitive. This allows them to be
stored in non-sensitive locations or even checked into repositories, as
long as the client secret file itself remains protected.