Don't block loading graphs on all user graphs being listed (#6681)
This should improve latency on loading every graph. Previously, we could
never load a graph until we had listed all of the user's graphs. But,
really that's only needed for the homepage, and since it can be slow, we
don't want to do that. We previously needed the list of graphs because
that's how we tracked ownership and versioning, so now we get that data
on a per-graph basis instead, when we load it.