Drop user-perceived latency for creating new graphs from ~5s to ~100ms (#6699)
This drops the user-perceived latency in creating a new graph from
around 5 seconds to around 100 milliseconds.

It works like this:

- The create method now uses
https://developers.google.com/workspace/drive/api/reference/rest/v3/files/generateIds
to get a file ID from Drive in the quickest way possible, and returns it
right away. From the caller's perspective, the graph is ready to display
and edit at this time.

- Meanwhile, the rest of the create operation is backgrounded, and the
initial blank graph descriptor is stored in memory. The full create
operation could take up to 5 seconds, because it has to serially find
the Drive folder ID, and then upload the data, and both of those calls
are quite slow.

- A subsequent load() operation for that URL will return that blank
graph directly from memory, and a subsequent save() or delete() will
block on that initial create completing.

- Once the background create is completed, and the first save() has
finished, everything works as it did before.