[unified-server] Create server-side flags for client-side env vars (#6702)
Set them in the client deployment config.

This doesn't do anything on its own, but sets up further work to replace
the VITE_ environment variables with their runtime counterparts. The
goal is to get rid of all build-time variables, so that the same image
can be deployed to multiple environments.