[cleanup] Remove connections list API (#6810)
In practice, we were only calling the connections/list API to find out
the OAuth client ID. Now we just pass the client ID directly into the
browser's static client deployment config object. Note this is possible
because we used to support multiple OAuth connections, but for a long
time now we have only supported 1.