Fix broken ability to view featured Opals while signed-out (#6835)
https://github.com/breadboard-ai/breadboard/pull/6796 broke the ability
to view featured Opals while signed-out, because it changed the behavior
of the geo-check in a way that would force the user to sign-in. The
intended design is that the geo-check doesn't run until you are
signed-in. Then, an otherwise unrelated bug related to how ui state
class gets initialized lazily caused the whole page to crash, instead of
the sign-in prompt to render. Both issues are now fixed.