[signin] Simplify and unify sign-in logic and make more amenable for opal shell (#6846)
- Simplified the internal API for sign-in. Just call
signinAdapter.signIn and wait for a result. This handles both opening
the window, and listening for the result. Previously those were two
separate steps, with the current nonce stored as class-state, so it was
in theory possible to initiate multiple conflicting sign-ins.

- We now first open an `about:blank` popup, and then immediately
navigate to the sign-in URL (which is possible because `about:blank` is
same-origin). The advantage of this is that it lets us open a window
synchronously in response to an event (the only way we are allowed to),
_before_ we know what the real sign-in URL is. This is what allows the
API to be simpler.

- We now always open a small floating window, instead of sometimes
opening a tab. We used to open a window from the signed-out graph page,
now we also do it from the landing page.