Emit ES2022 instead of ES2020 (#6927)
- Most notably, this means we'll emit real #private class fields instead
of the transpiled version.

- This fixes a very weird bug at
https://github.com/breadboard-ai/breadboard/blob/1d130e071ce63a43287a1d14ddc7e9af33d7a588/packages/shared-ui/src/flow-gen/app-catalyst.ts#L39
where a real #private class field works fine with comlink, but a
transpiled one leads to a "Cannot read private member from an object
whose class did not declare it" error later on. I do not understand
exactly why this bug occurs, but either using a real `#privateField`, or
a `private _regularField`, fixes it.

- Output will be a bit smaller.

See
https://caniuse.com/?feats=mdn-javascript_builtins_array_at%2Cmdn-javascript_builtins_regexp_hasindices%2Cmdn-javascript_builtins_object_hasown%2Cmdn-javascript_builtins_error_cause%2Cmdn-javascript_operators_await_top_level%2Cmdn-javascript_classes_private_class_fields%2Cmdn-javascript_classes_private_class_methods%2Cmdn-javascript_classes_static_class_fields%2Cmdn-javascript_classes_static_initialization_blocks
for browser support matrix of the ES2022 features.