Add consent manager for popups and get_webpage (#6946)
Introduces a `ConsentManager` to request consent before performing
potentially sensitive actions, like accessing a webpage or opening
popups.

* Adds`ConsentManager`, which manages querying, requesting, and storing
allow/deny consent for specific consent types (currently
`GET_ANY_WEBPAGE`, `OPEN_WEBPAGE`, but should also be usable for MCP) in
indexdb, keyed against the graphId, consent type, and consent scope
(which can vary per type). Methods:
* `queryConsent(request: ConsentRequest, askUsingUiType: ConsentUIType):
Promise<boolean | undefined>;`
* `setConsent(request: ConsentRequest, allow: boolean): Promise<void>;`
  * `revokeConsent(request: ConsentRequest): Promise<void>;`
* `getAllConsentsByType(type: ConsentRequest['type']):
Promise<ConsentRequest[]>;`
  * `clearAllConsents(): Promise<void>;`

* Adds `<bb-consent-modal>` element that takes a
`ConsentRequestWithCallback` and renders a modal popup to ask for
consent for the requested consent type/scope from the user, for when the
UI type is `ConsentUIType.MODAL`
* Adds a `#renderConsent` method to the app template that takes a
`ConsentRequestWithCallback` and renders a an in-app screen to ask for
consent for the requested consent type/scope from the user, for when the
UI type is `ConsentUIType.IN_APP`
* Adds `consentRequests: SignalArray<ConsentRequestWithCallback>` to UI
state (for `ConsentUIType.MODAL`) and app state (for
`ConsentUIType.IN_APP`), which `ConsentManager` uses to coordinate
opening the consent screen and getting the response
* Plumbs `ConsentManager` through to board event handler and output app
template
* Modifies the HTML output renderer to inject a script that patches
`window.open` and catches <a> clicks and posts back to the main page to
request Opal to open the popup