[shell] Handle shell reconnection for full page navigations (#7012)
We only do a full navigation in one place. On initial load, if the user
is signed-out, we do a full reload redirect to /landing/. Before this
PR, that would break the shell integration, because the shell did not
know how to re-negotiate a MessageChannel when the guest reloads. Now it
does.